fix: import from a renamed script
